refactor: move metrics logic to own file
